feat(fleet movement):
- Created PathEntry class to save important path information for future computations
- Paths calculated by the Dijkstra's algorithm will be saved and dynamically called to save computation time
- Implemented debug mode to display useful information about some game aspects
- travel path is now computed from the travel path by adding some intermediate points between islands. each travel points corresponds to one in-game period
- did some code refactoring, added documentation and comments

+++ src/main/java/de/uniks/stp24/service/Constants.java
@@ -267,8 +267,8 @@ public class Constants {
             NOT_BUILT
     }
 
-    public static final int ISLAND_HEIGHT = 162;
-    public static final int ISLAND_WIDTH = 70;
+    public static final double ISLAND_HEIGHT = 162;
+    public static final double ISLAND_WIDTH = 70;
     public static final int FLEET_FROM_ISLAND_DISTANCE = 40;
     public static final int FLEET_HW = 30;
